refactor(test): Refactored test packages
