+++ src/test/java/org/texttechnologylab/DockerUnifiedUIMAInterface/lua_process/TestSpacyEos.java
@@ -1,4 +1,4 @@
-package org.texttechnologylab.duui.lua_process;
+package org.texttechnologylab.DockerUnifiedUIMAInterface.lua_process;
 
 import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
 import org.apache.uima.cas.CASException;
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.texttechnologylab.DockerTestContainerManager;
+import org.texttechnologylab.DockerUnifiedUIMAInterface.test.DockerTestContainerManager;
 import org.texttechnologylab.DockerUnifiedUIMAInterface.DUUIComposer;
 import org.texttechnologylab.DockerUnifiedUIMAInterface.driver.DUUIRemoteDriver;
 import org.texttechnologylab.DockerUnifiedUIMAInterface.lua.DUUILuaContext;
